Make position indices and map dimensions long

+++ src/lazarus/SquareGridMap.cpp
@@ -14,7 +14,7 @@ void __lz::throw_out_of_bounds_exception(const Position2D &pos)
     throw __lz::LazarusException(msg.str());
 }
 
-Position2D::Position2D(int x, int y)
+Position2D::Position2D(long x, long y)
     : x(x)
     , y(y)
 {
@@ -37,7 +37,7 @@ bool Position2D::operator<(const Position2D &other) const
     return y < other.y;
 }
 
-SquareGridMap::SquareGridMap(unsigned width, unsigned height, bool diagonals)
+SquareGridMap::SquareGridMap(unsigned long width, unsigned long height, bool diagonals)
     : diagonals(diagonals)
     , width(width)
     , height(height)
@@ -48,12 +48,12 @@ SquareGridMap::SquareGridMap(unsigned width, unsigned height, bool diagonals)
         throw __lz::LazarusException("SquareGridMap width and height must be positive.");
 }
 
-unsigned SquareGridMap::get_width() const
+unsigned long SquareGridMap::get_width() const
 {
     return width;
 }
 
-unsigned SquareGridMap::get_height() const
+unsigned long SquareGridMap::get_height() const
 {
     return height;
 }
@@ -65,7 +65,7 @@ bool SquareGridMap::is_walkable(const Position2D &pos) const
     return costs[pos.y * width + pos.x] >= 0.;
 }
 
-bool SquareGridMap::is_walkable(int x, int y) const
+bool SquareGridMap::is_walkable(long x, long y) const
 {
     return is_walkable(Position2D(x, y));
 }
@@ -77,7 +77,7 @@ bool SquareGridMap::is_transparent(const Position2D &pos) const
     return transparencies[pos.y * width + pos.x];
 }
 
-bool SquareGridMap::is_transparent(int x, int y) const
+bool SquareGridMap::is_transparent(long x, long y) const
 {
     return is_transparent(Position2D(x, y));
 }
@@ -87,7 +87,7 @@ bool SquareGridMap::is_out_of_bounds(const Position2D &pos) const
     return is_out_of_bounds(pos.x, pos.y);
 }
 
-bool SquareGridMap::is_out_of_bounds(int x, int y) const
+bool SquareGridMap::is_out_of_bounds(long x, long y) const
 {
     return x < 0 || y < 0 || x >= width || y >= height;
 }
@@ -111,7 +111,7 @@ float SquareGridMap::get_cost(const Position2D &pos) const
     return costs[pos.y * width + pos.x];
 }
 
-float SquareGridMap::get_cost(int x, int y) const
+float SquareGridMap::get_cost(long x, long y) const
 {
     return get_cost(Position2D(x, y));
 }
@@ -123,7 +123,7 @@ std::vector<Position2D> SquareGridMap::neighbours(const Position2D &pos) const
         __lz::throw_out_of_bounds_exception(pos);
 
     std::vector<Position2D> result;
-    int x = pos.x, y = pos.y;
+    long x = pos.x, y = pos.y;
 
     std::array<Position2D, 4> neighbour_positions{
         Position2D(x - 1, y),
@@ -155,7 +155,7 @@ std::vector<Position2D> SquareGridMap::neighbours(const Position2D &pos) const
     return result;
 }
 
-std::vector<Position2D> SquareGridMap::neighbours(int x, int y) const
+std::vector<Position2D> SquareGridMap::neighbours(long x, long y) const
 {
     return neighbours(Position2D(x, y));
 }
@@ -171,7 +171,7 @@ void SquareGridMap::set_cost(const Position2D &pos, float cost)
     costs[pos.y * width + pos.x] = cost;
 }
 
-void SquareGridMap::set_cost(int x, int y, float cost)
+void SquareGridMap::set_cost(long x, long y, float cost)
 {
     set_cost(Position2D(x, y), cost);
 }
@@ -183,7 +183,7 @@ void SquareGridMap::set_walkable(const Position2D &pos, bool walkable)
     set_cost(pos, walkable ? 1. : -1.);
 }
 
-void SquareGridMap::set_walkable(int x, int y, bool walkable)
+void SquareGridMap::set_walkable(long x, long y, bool walkable)
 {
     set_walkable(Position2D(x, y), walkable);
 }
@@ -199,7 +199,7 @@ void SquareGridMap::set_transparency(const Position2D &pos, bool transparent)
     transparencies[pos.y * width + pos.x] = transparent;
 }
 
-void SquareGridMap::set_transparency(int x, int y, bool transparent)
+void SquareGridMap::set_transparency(long x, long y, bool transparent)
 {
     set_transparency(Position2D(x, y), transparent);
 }
@@ -208,9 +208,9 @@ void SquareGridMap::carve_room(const Position2D &top_left,
                                const Position2D &bottom_right,
                                float cost)
 {
-    for (int x = top_left.x; x <= bottom_right.x; ++x)
+    for (long x = top_left.x; x <= bottom_right.x; ++x)
     {
-        for (int y = top_left.y; y <= bottom_right.y; ++y)
+        for (long y = top_left.y; y <= bottom_right.y; ++y)
         {
             Position2D pos(x, y);
             set_cost(pos, cost);
